chore: rename mediator package to cqrs
